Rename codepages module to encoding, fix typo
